requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-SSLify==0.1.5
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
six==1.12.0
slackclient==2.0.0
Werkzeug==0.15.2
yarl==1.3.0
